+++ src/OctoshiftCLI.Tests/ado2gh/Commands/GenerateScriptCommandTests.cs
@@ -9,6 +9,7 @@ using Moq;
 using Octoshift.Models;
 using OctoshiftCLI.AdoToGithub;
 using OctoshiftCLI.AdoToGithub.Commands;
+using OctoshiftCLI.AdoToGithub.Handlers;
 using OctoshiftCLI.Contracts;
 using OctoshiftCLI.Extensions;
 using Xunit;
