+++ src/OctoshiftCLI.Tests/ado2gh/Commands/IntegrateBoardsCommandTests.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Moq;
 using OctoshiftCLI.AdoToGithub;
 using OctoshiftCLI.AdoToGithub.Commands;
+using OctoshiftCLI.AdoToGithub.Handlers;
 using Xunit;
 
 namespace OctoshiftCLI.Tests.AdoToGithub.Commands
