+++ src/OctoshiftCLI.Tests/ado2gh/Commands/InventoryReportCommandTests.cs
@@ -3,6 +3,7 @@ using FluentAssertions;
 using Moq;
 using OctoshiftCLI.AdoToGithub;
 using OctoshiftCLI.AdoToGithub.Commands;
+using OctoshiftCLI.AdoToGithub.Handlers;
 using Xunit;
 
 namespace OctoshiftCLI.Tests.AdoToGithub.Commands
