+++ src/ado2gh/Handlers/IntegrateBoardsCommandHandler.cs
@@ -2,8 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using OctoshiftCLI.AdoToGithub.Commands;
 
-namespace OctoshiftCLI.AdoToGithub.Commands
+namespace OctoshiftCLI.AdoToGithub.Handlers
 {
     public class IntegrateBoardsCommandHandler
     {
