+++ src/ado2gh/Handlers/ShareServiceConnectionCommandHandler.cs
@@ -1,7 +1,8 @@
 ﻿using System;
 using System.Threading.Tasks;
+using OctoshiftCLI.AdoToGithub.Commands;
 
-namespace OctoshiftCLI.AdoToGithub.Commands
+namespace OctoshiftCLI.AdoToGithub.Handlers
 {
     public class ShareServiceConnectionCommandHandler
     {
